Initialize the todos variable with objects
The app stores the text and state of a todo. The state is extra information that marks if the todo is completed or incomplete.

There were two options to structure the data: Arrays in the array or objects in the array. I used objects.

To come up with the decision I checked the pros and cons of using each option. Using objects has a strong advantage, which is that we get to name the properties. This is relevant when trying to access the data.

This change - rewriting the variable to use objects instead of strings - means we also have to change the functions to add and edit a todo.