Rewrite add function to add objects
The app adds todos as objects.

The add function was changed so that the argument passed in is saved to a property on an object.

The default status for the completed property is set to false, because when the user adds a todo to the list, the todo is not going to be completed yet.